refactor: MoveResolver를 MovePolicy로 인터페이스 추출후 RandomMovePolicy로 구현
