+++ src/main/java/racingcar/model/MovePolicy.java
@@ -1,0 +1,5 @@
+package racingcar.model;
+
+public interface MovePolicy {
+    boolean isMoveAble();
+}
