+++ src/main/java/racingcar/model/RandomMovePolicy.java
@@ -1,8 +1,8 @@
-package racingcar.util;
+package racingcar.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-public class MoveResolver {
+public class RandomMovePolicy implements MovePolicy{
     private final int MOVING_FORWARD = 4;
     private final int RANDOM_LEFT_BOUND = 0;
     private final int RANDOM_RIGHT_BOUND = 9;
